cpufreq: fallback to interactive if governor is not found
Lots of custom ROMs nowadays are inlining custom kernels and adding their own tweaks via init scripts; this may include a governor profile. If the user flashes another kernel that does not have the governor requested by the boot script, it will be forced to fallback to the default governor. In this kernel (and pretty much every other Android kernel), the default governor is performance, which is then changed to interactive on boot. So, if the governor requested during boot is not found, we'd be stuck with performance. For user experience, let's try falling back to interactive before erroring out.

Couple of notes: This is a quick fix; I would like to research a way to check which governor is currently active, as this is really only intended for the use case where the governor is stuck on performance without the user's consent. The only reason this isn't a high priority is kernel tweaking apps will only show possible governor options so this will never trigger under normal use cases. Lastly, if you have a default governor other than interactive, change it after applying this commit.

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>